+++ attrib/Class.py
@@ -21,6 +21,7 @@ class Class(base.DataAttribute):
         "Python 3.9"
     )
     base.VERSION.changed("1.8.0", "Replaced Legacy format strings by f-strings in `attrib.Class` ")
+    base.VERSION.changed("1.9.0", "Switched to Google docstring style in `attrib.Class` ")
 
     def __init__(self, expected_type: type, severity: int = 2) -> None:
         """