+++ attrib/Equals.py
@@ -15,6 +15,7 @@ class Equals(base.DataAttribute):
     base.VERSION.changed("1.7.0", "`attrib.Equals` got new base class `base.DataAttribute` ")
     base.VERSION.added("1.7.0", "Type hints to `attrib.Equals` ")
     base.VERSION.changed("1.8.0", "Replaced Legacy format strings by f-strings in `attrib.Equals` ")
+    base.VERSION.changed("1.9.0", "Switched to Google docstring style in `attrib.Equals` ")
 
     def __init__(self, value: typing.Any, severity: int = 1) -> None:
         """