+++ attrib/Ontology.py
@@ -13,6 +13,7 @@ class Ontology(base.DataAttribute):
     base.VERSION.changed("1.7.0", "`attrib.Ontology` got new base class `base.DataAttribute` ")
     base.VERSION.added("1.7.0", "Type hints to `attrib.Ontology` ")
     base.VERSION.changed("1.7.0", "Removed deactivated code in `attrib.Ontology` ")
+    base.VERSION.changed("1.9.0", "Switched to Google docstring style in `attrib.Ontology` ")
 
     def __init__(self, iri: str) -> None:
         """