+++ base/Store.py
@@ -11,6 +11,7 @@ class Store:
     base.VERSION.added("1.4.1", "Changelog in `base.Store` ")
     base.VERSION.changed("1.5.3", "`base.Store` changelog uses markdown for code elements")
     base.VERSION.added("1.7.0", "Type hints to `base.Store` ")
+    base.VERSION.changed("1.9.0", "Switched to Google docstring style in `base.Store` ")
 
     def close(self) -> None:
         """