+++ components/CsvReader.py
@@ -25,6 +25,7 @@ class CsvReader(base.Component):
     base.VERSION.changed("1.5.3", "`components.CsvReader` changelog uses markdown for code elements")
     base.VERSION.added("1.7.0", "Type hints to `components.CsvReader` ")
     base.VERSION.changed("1.7.0", "Harmonized init signature of `components.CsvReader` with base class")
+    base.VERSION.changed("1.9.0", "Switched to Google docstring style in `component.CsvReader` ")
 
     def __init__(self, name: str, default_observer: base.Observer, default_store: typing.Optional[base.Store]) -> None:
         """