+++ components/ReportingDistribution.py
@@ -31,6 +31,7 @@ class ReportingDistribution(base.Component):
     base.VERSION.changed("1.5.3", "`components.ReportingDistribution` changelog uses markdown for code elements")
     base.VERSION.added("1.7.0", "Type hints to `components.ReportingDistribution` ")
     base.VERSION.changed("1.7.0", "Harmonized init signature of `components.ReportingDistribution` with base class")
+    base.VERSION.changed("1.9.0", "Switched to Google docstring style in `component.ReportingDistribution` ")
 
     def __init__(self, name: str, default_observer: base.Observer, default_store: typing.Optional[base.Store]) -> None:
         """