+++ components/__init__.py
@@ -1,6 +1,4 @@
-"""
-Imports of the `components` module.
-"""
+"""Imports of the `components` module."""
 from .CsvReader import *
 from .DeleteFolder import *
 from .DepositionToPecSoil import *
@@ -27,3 +25,4 @@ base.VERSION.added("1.4.1", "Changelog in `components` ")
 base.VERSION.changed("1.4.1", "Removed `components.VERSION` ")
 base.VERSION.changed("1.4.9", "`components` imports updated to reflect renamed components")
 base.VERSION.changed("1.6.1", "`components` imports updated to reflect renamed components")
+base.VERSION.changed("1.9.0", "Switched to Google docstring style in `component.__init__` ")