+++ extension/__init__.py
@@ -11,3 +11,4 @@ base.VERSION.added("1.1.6", "Changelog through `VersionInfo` class in `extension
 base.VERSION.added("1.4.1", "Changelog in `extension.__init__` ")
 base.VERSION.changed("1.4.1", "Removed `extension.VERSION` ")
 base.VERSION.changed("1.5.3", "`extension` changelog uses markdown for code elements")
+base.VERSION.changed("1.9.0", "Renamed `extend` namespace to `extension` ")