+++ stores/SqlLiteStore.py
@@ -36,6 +36,7 @@ class SqlLiteStore(base.Store):
     base.VERSION.added("1.7.0", "Type hints to `stores.SqlLiteStore` ")
     base.VERSION.fixed("1.7.0", "Check for slices containing steps in `stores.SqlLiteStore` ")
     base.VERSION.changed("1.8.0", "Replaced Legacy format strings by f-strings in `stores.SqlLiteStore` ")
+    base.VERSION.changed("1.9.0", "Switched to Google docstring style in `stores.SqlLiteStore` ")
 
     def __init__(self, file_path: str, observer: base.Observer, create: bool = True) -> None:
         """